requirements.txt
aiogram
asyncio
requests
python-dotenv 
